modifier effacer et ajouter maladie enfant fait, suppression propre en bdd, sécurité ajouté

+++ src/Controller/DiseaseController.php
@@ -5,6 +5,7 @@ namespace App\Controller;
 use App\Entity\Disease;
 use App\Entity\Child;
 use App\Form\DiseaseType;
+use phpDocumentor\Reflection\Types\Null_;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
@@ -51,7 +52,7 @@ class DiseaseController extends AbstractController
                     return $this->redirectToRoute('profil');
                 }
 
-                return $this->render('disease/new.html.twig', [
+                return $this->render('disease/index.html.twig', [
                     'disease' => $disease,
                     'form' => $form->createView(),
                 ]);
@@ -68,36 +69,60 @@ class DiseaseController extends AbstractController
      * @Route("/{id}/edit", name="disease_edit", methods={"GET","POST"})
      */
     public
-    function edit(Request $request, Child $child, Disease $disease): Response
+    function edit(Request $request, Disease $disease): Response
     {
-        $child->getDiseases();
-        $form = $this->createForm(DiseaseType::class, $disease);
-        $form->handleRequest($request);
+        $enfants = $this->getUser()->getChildren();
+        $i = 0;
+        $temp = false;
 
-        if ($form->isSubmitted() && $form->isValid()) {
-            $this->getDoctrine()->getManager()->flush();
+        while ($enfants[$i] != Null) {
+            $o = 0;
+            $maladie = $enfants[$i]->getDiseases();
 
-            return $this->redirectToRoute('profil');
+            while ($maladie[$o] != Null) {
+                if ($disease->getId() == $maladie[$o]->getId()) {
+                    $temp = true;
+                }
+                $o++;
+            }
+            $i++;
         }
 
-        return $this->render('disease/edit.html.twig', [
-            'disease' => $disease,
-            'form' => $form->createView(),
-        ]);
+        if ($temp == true){
+            $form = $this->createForm(DiseaseType::class, $disease);
+            $form->handleRequest($request);
+
+            if ($form->isSubmitted() && $form->isValid()) {
+                $this->getDoctrine()->getManager()->flush();
+
+                return $this->redirectToRoute('profil');
+            }
+
+            return $this->render('disease/edit.html.twig', [
+                'disease' => $disease,
+                'form' => $form->createView(),
+            ]);
+        } else {
+            return $this->render('403/403.html.twig', ['erreur' => 'ACCES FORBIDEN']);
+        }
     }
 
     /**
      * @Route("/delete/{id}", name="disease_delete", methods={"DELETE"})
      */
     public
-    function delete(Request $request, Disease $disease, Child $child): Response
+    function delete(Request $request, Disease $disease, AuthorizationCheckerInterface $authChecker): Response
     {
-        if ($this->isCsrfTokenValid('delete' . $disease->getId(), $request->request->get('_token'))) {
-            $entityManager = $this->getDoctrine()->getManager();
-            $entityManager->remove($disease);
-            $entityManager->flush();
-        }
+        if (true === $authChecker->isGranted('ROLE_PARENT')) {
+            if ($this->isCsrfTokenValid('delete' . $disease->getId(), $request->request->get('_token'))) {
+                $entityManager = $this->getDoctrine()->getManager();
+                $entityManager->remove($disease);
+                $entityManager->flush();
+            }
 
-        return $this->redirectToRoute('profil');
+            return $this->redirectToRoute('profil');
+        } else {
+            return $this->render('403/403.html.twig', ['erreur' => 'ACCES FORBIDEN']);
+        }
     }
 }
